requirements.txt
botcity-framework-web>=0.3.1
botcity-maestro-sdk>=0.1.3
lxml
botcity-utils>=0.1.2
pandas
requests
pyqt5
botcity-crawler-plugin